ecommerce/requirements.txt
absl-py==1.0.0
arabic-reshaper==2.1.3
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.4.0
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
h5py==3.6.0
html5lib==1.1
httplib2==0.19.1
huggingface-hub==0.1.2
idna==3.2
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keybert==0.5.0
libclang==12.0.0
Markdown==3.3.6
nltk==3.6.5
numpy==1.21.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.3.2
protobuf==3.18.0
psycopg2==2.9
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytube==11.0.1
pytz==2021.1
PyYAML==6.0
razorpay==1.2.0
regex==2021.11.10
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.14.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.5
typing_extensions==4.0.0
uritemplate==3.0.1
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
xhtml2pdf==0.2.5
XlsxWriter==3.0.1
zipp==3.6.0
